closes #1, Implement a basic system for triggering plot flags
Plot triggers can be attached to player choices, and used to filter NPC responses later in the game. Added the ability to link an NPC dialog from a different part of the tree structure and carry on the conversation from that point. Plots are managed across the game and across scenes by a plot manager singleton.